refactor: 활동모델 클래스명 변경
