+++ app/src/main/java/com/teamwss/websoso/ui/activityDetail/adapter/ActivityDetailViewHolder.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.teamwss.websoso.databinding.ItemMyActivityBinding
-import com.teamwss.websoso.ui.main.myPage.myActivity.model.ActivityModel
+import com.teamwss.websoso.ui.main.myPage.myActivity.model.ActivitiesModel.ActivityModel
 
 class ActivityDetailViewHolder(private val binding: ItemMyActivityBinding) :
     RecyclerView.ViewHolder(binding.root) {
